# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Social Circle API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: contacts
    description: Contact management
  - name: interactions
    description: Interaction log
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /contacts:
    get:
      operationId: listContacts
      tags: [contacts]
      summary: List all contacts
      parameters:
        - name: tier
          in: query
          required: false
          schema:
            type: string
            enum: [core, monthly, yearly]
        - name: overdue
          in: query
          required: false
          schema:
            type: boolean
        - name: archived
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: List of contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contact"
    post:
      operationId: createContact
      tags: [contacts]
      summary: Create a contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContactBody"
      responses:
        "201":
          description: Created contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"

  /contacts/stats:
    get:
      operationId: getContactStats
      tags: [contacts]
      summary: Get summary stats across all tiers
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactStats"

  /contacts/due:
    get:
      operationId: getDueContacts
      tags: [contacts]
      summary: Get contacts due or overdue for follow-up, sorted by urgency
      responses:
        "200":
          description: Contacts due for follow-up
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DueContact"

  /contacts/{id}:
    get:
      operationId: getContact
      tags: [contacts]
      summary: Get a contact by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateContact
      tags: [contacts]
      summary: Update a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContactBody"
      responses:
        "200":
          description: Updated contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteContact
      tags: [contacts]
      summary: Delete a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /contacts/{id}/touch:
    post:
      operationId: touchContact
      tags: [contacts]
      summary: Mark contact as reached out today, recalculates next contact date
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"

  /contacts/{id}/archive:
    post:
      operationId: archiveContact
      tags: [contacts]
      summary: Archive a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Archived contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"

  /contacts/{id}/unarchive:
    post:
      operationId: unarchiveContact
      tags: [contacts]
      summary: Unarchive a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Unarchived contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"

  /contacts/{id}/interactions:
    get:
      operationId: listInteractions
      tags: [interactions]
      summary: List interactions for a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of interactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Interaction"
    post:
      operationId: createInteraction
      tags: [interactions]
      summary: Create an interaction for a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInteractionBody"
      responses:
        "201":
          description: Created interaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Interaction"

  /contacts/{id}/interactions/{interactionId}:
    delete:
      operationId: deleteInteraction
      tags: [interactions]
      summary: Delete an interaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: interactionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /contacts/{id}/calendar.ics:
    get:
      operationId: getContactCalendar
      tags: [contacts]
      summary: Get iCal (.ics) reminder for a contact's next follow-up
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: iCal file
          content:
            text/calendar:
              schema:
                type: string

  /calendar/token:
    get:
      operationId: getCalendarToken
      tags: [contacts]
      summary: Get the ICS subscription feed token and URL
      responses:
        "200":
          description: Calendar feed token and subscription URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarToken"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    Contact:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        tier:
          type: string
          enum: [core, monthly, yearly]
        intervalDays:
          type: integer
          nullable: true
          description: Number of days between contacts. Overrides the tier default.
        relationshipType:
          type: string
        lastContactDate:
          type: string
          format: date
          nullable: true
        nextContactDate:
          type: string
          format: date
          nullable: true
        notes:
          type: string
          nullable: true
        birthday:
          type: string
          format: date
          nullable: true
        tags:
          type: array
          items:
            type: string
        streak:
          type: integer
          description: Consecutive on-time touch count
        archivedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - tier
        - intervalDays
        - relationshipType
        - lastContactDate
        - nextContactDate
        - notes
        - birthday
        - tags
        - streak
        - archivedAt
        - createdAt
        - updatedAt

    CreateContactBody:
      type: object
      properties:
        name:
          type: string
        tier:
          type: string
          enum: [core, monthly, yearly]
        intervalDays:
          type: integer
          nullable: true
          description: Number of days between contacts. Defaults to tier standard if omitted.
        relationshipType:
          type: string
        lastContactDate:
          type: string
          format: date
          nullable: true
        notes:
          type: string
          nullable: true
        birthday:
          type: string
          format: date
          nullable: true
        tags:
          type: array
          items:
            type: string
      required:
        - name
        - tier
        - relationshipType

    UpdateContactBody:
      type: object
      properties:
        name:
          type: string
        tier:
          type: string
          enum: [core, monthly, yearly]
        intervalDays:
          type: integer
          nullable: true
          description: Number of days between contacts.
        relationshipType:
          type: string
        lastContactDate:
          type: string
          format: date
          nullable: true
        nextContactDate:
          type: string
          format: date
          nullable: true
        notes:
          type: string
          nullable: true
        birthday:
          type: string
          format: date
          nullable: true
        tags:
          type: array
          items:
            type: string
        archivedAt:
          type: string
          format: date-time
          nullable: true

    ContactStats:
      type: object
      properties:
        total:
          type: integer
        core:
          type: integer
        monthly:
          type: integer
        yearly:
          type: integer
        overdueCount:
          type: integer
        dueThisWeek:
          type: integer
      required:
        - total
        - core
        - monthly
        - yearly
        - overdueCount
        - dueThisWeek

    CalendarToken:
      type: object
      properties:
        token:
          type: string
        feedUrl:
          type: string
      required:
        - token
        - feedUrl

    DueContact:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        tier:
          type: string
          enum: [core, monthly, yearly]
        relationshipType:
          type: string
        lastContactDate:
          type: string
          format: date
          nullable: true
        nextContactDate:
          type: string
          format: date
          nullable: true
        notes:
          type: string
          nullable: true
        daysOverdue:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - tier
        - relationshipType
        - lastContactDate
        - nextContactDate
        - notes
        - daysOverdue
        - createdAt
        - updatedAt

    Interaction:
      type: object
      properties:
        id:
          type: integer
        contactId:
          type: integer
        date:
          type: string
          format: date
        type:
          type: string
          enum: [call, email, coffee, message, video_call, other]
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - contactId
        - date
        - type
        - notes
        - createdAt

    CreateInteractionBody:
      type: object
      properties:
        date:
          type: string
          format: date
        type:
          type: string
          enum: [call, email, coffee, message, video_call, other]
        notes:
          type: string
          nullable: true
      required:
        - date
        - type